% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genDNABin.R
\name{genDNABin}
\alias{genDNABin}
\title{Generate a DNA Bin}
\usage{
genDNABin(specimen_dataframe)
}
\arguments{
\item{specimen_dataframe}{A dataframe with specimen data created using querySpecData()}
}
\value{
A DNA bin object.
}
\description{
Create a DNA bin object using a specimen dataframe.
}
\examples{
# generate a DNA bin object using the taxon 'Antheraea polyphemus'
specdf_Anth <- querySpecData("Antheraea polyphemus")[1:10,]

DNABin_Anth <- genDNABin(specdf_Anth)
}
