% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenusRepr.R
\name{getGenusRepr}
\alias{getGenusRepr}
\title{Get Genus Representatives}
\usage{
getGenusRepr(specimen_dataframe)
}
\arguments{
\item{specimen_dataframe}{A specimen dataframe.}
}
\value{
A dataframe where each column pertains to a variable and each row an individual specimen.
}
\description{
Get one observation/row per genus from a specimen dataframe.
}
\examples{
# query data then extract one observation for each of the genera available
specdata <- querySpecData("Alouatta")

specdata <- getGenusRepr(specdata)
}
