% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpeciesRepr.R
\name{getSpeciesRepr}
\alias{getSpeciesRepr}
\title{Get Species Representatives}
\usage{
getSpeciesRepr(specimen_dataframe)
}
\arguments{
\item{specimen_dataframe}{A specimen dataframe.}
}
\value{
A dataframe where each column pertains to a variable and each row an individual specimen.
}
\description{
Get one observation/row per species from a specimen dataframe.
}
\examples{
# query data then extract one observation for each of the species available
specdata <- PhyInsight::querySpecData("Alouatta")

specdata <- getSpeciesRepr(specdata)
}
