% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querySpecData.R
\name{querySpecData}
\alias{querySpecData}
\title{Query Specimen Data}
\usage{
querySpecData(taxonName)
}
\arguments{
\item{taxonName}{A character string that is the scientific name of a chosen taxon.}
}
\value{
A dataframe where each column pertains to a variable and each row an individual specimen.
}
\description{
Get specimen and DNA sequence data from the Bold Systems database returned as a dataframe.
}
\examples{
# generate a specimen data frame using the taxon 'Antheraea polyphemus'
specdf_Anth <- querySpecData("Antheraea polyphemus")[1:10,]
}
