% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savePhytree.R
\name{savePhytree}
\alias{savePhytree}
\title{Save a Phylogenetic Tree Object}
\usage{
savePhytree(
  phyloTree,
  specimen_dataframe,
  tree_file_name = "Phylogenetic Tree.pdf",
  open_PDF = T,
  label_offset = 5.5e-05,
  label_size = 0.3
)
}
\arguments{
\item{phyloTree}{A phylo tree object}

\item{specimen_dataframe}{A specimen dataframe.}

\item{tree_file_name}{An optional character string to name the PDF file.}

\item{open_PDF}{A logical value to state whether to open the PDF file.}

\item{label_offset}{A numerical value to state the label offset distance.}
}
\value{
A PDF file.
}
\description{
Create a phylogenetic tree PDF using a phylo tree object and a specimen dataframe.
}
\examples{
# create and plot a phylo tree
specdf_Anth <- querySpecData("Antheraea polyphemus")[1:10,]

DNABin_Anth <- genDNABin(specdf_Anth)

DNAStringset_Anth <- genDNAStringSet(DNABin_Anth)

DNAStringSet_Anth_manipulated <- ManipStringSet(DNAStringset_Anth)

Phytree_Anth <- genPhytree(DNAStringSet_Anth_manipulated)

savePhytree(Phytree_Anth, specdf_Anth, "Anth_phylo_tree", 0.000055)
}
